passing card to detail and able to retrieve detail api

+++ Foozy/Foozy/Models/Decodable/BusinessDetailData.swift
@@ -34,8 +34,8 @@ struct Open: Decodable {
 }
 
 struct Coordinates: Decodable {
-    let latitude: Float
-    let longitude: Float
+    let latitude: Double
+    let longitude: Double
 }
 
 struct Location: Decodable {
